Slightly optimise shuffling loop
This should result in minor improvements to the speed at which the next track is selected when large playlists are being shuffled and most of the tracks have already been played.

This is due to the negation in the original statement being a separate logic operation, which is now avoided. 

Was this worth the effort? I don't know, but it has been done.